drop "MIT License" line from LICENSE
I was importing this crate into Google's source control system. The
automation at Google doesn't like the "MIT License" line at the
beginning of the LICENSE-MIT file. AFAICT, the automation is correct
in thinking that the line is not supposed to be there.

For reference, I've made the same change in gitoxide:
https://github.com/Byron/gitoxide/pull/1097